apis/detect_files/requirements.txt
Flask==1.1.1
flask-restful==0.3.7
Pillow==6.2.0
